♻️  Button 네이밍 및 주석을 유연하게 작성되도록 변경
